move SaveData class to its own file for designer compliance
